requirements.txt
chromadb>=0.4.22
python-dotenv>=1.0.0
numpy>=1.24.0